Moves component to blocks
